refactor(card): remove unused methods + revert import dependency
