+++ app/core/Engine/controllers/card-controller/utils/moneyAccountCardToken.test.ts
@@ -1,4 +1,7 @@
-import { FundingStatus, DelegationSettingsResponse } from '../types';
+import {
+  FundingStatus,
+  DelegationSettingsResponse,
+} from '../../../../../components/UI/Card/types';
 import {
   hasMoneyAccountCardRequirements,
   resolveMoneyAccountCardToken,
