+++ app/core/Engine/controllers/card-controller/utils/moneyAccountCardToken.ts
@@ -1,5 +1,3 @@
-import type { CaipChainId } from '@metamask/utils';
-import { SolScope } from '@metamask/keyring-api';
 import {
   CardFundingToken,
   DelegationSettingsResponse,
@@ -8,19 +6,6 @@ import {
 
 const MONEY_ACCOUNT_CARD_NETWORK = 'monad';
 const MONEY_ACCOUNT_CARD_TOKEN_SYMBOL = 'USDC';
-
-function getCaipChainId(
-  network: DelegationSettingsResponse['networks'][0],
-): CaipChainId {
-  if (network.network === 'solana') {
-    return SolScope.Mainnet as CaipChainId;
-  }
-  const chainIdStr = network.chainId;
-  const numericChainId = chainIdStr.startsWith('0x')
-    ? parseInt(chainIdStr, 16)
-    : parseInt(chainIdStr, 10);
-  return `eip155:${numericChainId}` as CaipChainId;
-}
 
 interface MoneyAccountCardRequirementsParams {
   isMoneyAccountEnabled: boolean;
@@ -63,7 +48,7 @@ export const resolveMoneyAccountCardToken = (
     symbol: MONEY_ACCOUNT_CARD_TOKEN_SYMBOL,
     name: MONEY_ACCOUNT_CARD_TOKEN_SYMBOL,
     decimals: token.decimals,
-    caipChainId: getCaipChainId(network),
+    caipChainId: 'eip155:143',
     walletAddress: undefined,
     fundingStatus: FundingStatus.NotEnabled,
     spendableBalance: '0',
